api_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
opencv-python==4.9.0.80
pydantic==2.6.4
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
ultralytics==8.1.37
typing_extensions==4.10.0

pydantic==2.6.4